llama2-pinecone-rag/requirements.txt

langchain
langchain-openai
langchain-pinecone
pinecone-client
datasets
tqdm
python-dotenv
pandas
